Fixed specs for all basic classes
