+++ .rspec
@@ -1,3 +1,4 @@
 --format documentation
 --color
 --require spec_helper
+--fail-fast
